make contact controllers/routes private

+++ controllers/contactController.js
@@ -3,16 +3,16 @@ const Contact = require('../models/contactModel');
 
 //@desc Get all contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 const getContacts = asyncHandler(async (req, res) => {
-    const contacts = await Contact.find({});
+    const contacts = await Contact.find({user_id : req.user.id});
     res.status(200).json(contacts);
 });
 
 
 //@desc Get One contact
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 const getOneContact = asyncHandler(async (req, res) => {
     const id = req.params._id;
     const contact = await Contact.findById(id);
@@ -28,7 +28,7 @@ const getOneContact = asyncHandler(async (req, res) => {
 
 //@desc create new contact
 //@route POST /api/contacts
-//@access public
+//@access private
 const createContact = asyncHandler(async (req, res) => {
     const { name, email, phone } = req.body;
     if (!name || !email || !phone) {
@@ -39,7 +39,8 @@ const createContact = asyncHandler(async (req, res) => {
     const contact = await Contact.create({
         name,
         email,
-        phone
+        phone,
+        user_id: req.user.id
     });
 
     res.status(201).json(contact);
@@ -48,7 +49,7 @@ const createContact = asyncHandler(async (req, res) => {
 
 //@desc Update one contact
 //@route PUT /api/contacts/:id
-//@access public
+//@access private
 const updateContact = asyncHandler(async (req, res) => {
     const id = req.params._id;
     const contact = await Contact.findById(id);
@@ -56,6 +57,11 @@ const updateContact = asyncHandler(async (req, res) => {
     if (!contact) {
         res.status(404);
         throw new Error('Contact Not Found!');
+    }
+
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error(`user don't have permission to update other user contacts!`);
     }
 
     const updatedContact = await Contact.findByIdAndUpdate(
@@ -70,7 +76,7 @@ const updateContact = asyncHandler(async (req, res) => {
 
 //@desc Delete one contact
 //@route DELETE /api/contacts/:id
-//@access public
+//@access private
 const deleteContact = asyncHandler(async (req, res) => {
     const id = req.params._id;
     const contact = await Contact.findById(id);
@@ -78,6 +84,11 @@ const deleteContact = asyncHandler(async (req, res) => {
     if (!contact) {
         res.status(404);
         throw new Error('Contact Not Found!');
+    }
+
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error(`user don't have permission to delete other user contacts!`);
     }
 
     const contacts = await Contact.findByIdAndDelete(id);
